Refactor FetchImpl
Move ReleaseDefinition and validateReleaseDefinition out of fetchImpl, so that
they can be used by the Release command as well.
