+++ packages/sfpowerscripts-cli/src/impl/release/ReleaseDefinitionInterface.ts
@@ -1,0 +1,6 @@
+export default interface ReleaseDefinition {
+  release: string,
+  artifacts: {
+    [p: string]: string
+  }
+}
